Fix clippy cast warning

+++ objc2/src/runtime.rs
@@ -416,7 +416,7 @@ impl Class {
 
     /// Returns the size of instances of self.
     pub fn instance_size(&self) -> usize {
-        unsafe { ffi::class_getInstanceSize(self.as_ptr()) as usize }
+        unsafe { ffi::class_getInstanceSize(self.as_ptr()) }
     }
 
     /// Returns a specified instance method for self, or [`None`] if self and
